Move FastModifiableObservableList to public class.
